cleaner

+++ app/database/migrations/2021_11_06_201718_create_cards_table.php
@@ -19,7 +19,7 @@ class CreateCardsTable extends Migration
             
             $table->string('type')->nullable();
             $table->string('name')->nullable();
-            $table->integer('number')->nullable();
+            $table->bigInteger('number')->nullable();
             $table->string('expirationDate')->nullable();
             
             $table->timestamps();
